refactor(calculator): String Calculator 패키지 위치 변경.
